+++ OPHD/UI/Core/RadioButtonGroup.cpp
@@ -56,7 +56,7 @@ void RadioButtonGroup::select(std::size_t index)
 
 void RadioButtonGroup::select(RadioButtonGroup::RadioButton& button)
 {
-	auto index = std::distance(mRadioButtons.data(), &button);
+	auto index = static_cast<std::size_t>(std::distance(mRadioButtons.data(), &button));
 	select(index);
 }
